Move util interfaces into ExternalInterfaces
