+++ compiler/src/iree/compiler/Dialect/Util/IR/CMakeLists.txt
@@ -12,25 +12,6 @@ iree_add_all_subdirs()
 
 iree_cc_library(
   NAME
-    ExternalModels
-  HDRS
-    "UtilExternalModels.h"
-  SRCS
-    "UtilExternalModels.cpp"
-  DEPS
-    ::IR
-    MLIRArithDialect
-    MLIRLinalgDialect
-    MLIRLinalgOpsIncGenLib
-    MLIRLinalgStructuredOpsIncGenLib
-    MLIRMLProgramDialect
-    MLIRTensorDialect
-    iree::compiler::Dialect::LinalgExt::IR
-  PUBLIC
-)
-
-iree_cc_library(
-  NAME
     IR
   HDRS
     "ClosureOpUtils.h"
