+++ compiler/src/iree/compiler/ExternalInterfaces/CMakeLists.txt
@@ -1,0 +1,34 @@
+################################################################################
+# Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
+# compiler/src/iree/compiler/Dialect/ExternalInterfaces/BUILD.bazel            #
+#                                                                              #
+# Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
+# CMake-only content.                                                          #
+#                                                                              #
+# To disable autogeneration for this file entirely, delete this header.        #
+################################################################################
+
+iree_add_all_subdirs()
+
+iree_cc_library(
+  NAME
+    ExternalModels
+  HDRS
+    "Interfaces.h"
+    "UtilExternalModels.h"
+  SRCS
+    "Interfaces.cpp"
+    "UtilExternalModels.cpp"
+  DEPS
+    MLIRArithDialect
+    MLIRLinalgDialect
+    MLIRLinalgOpsIncGenLib
+    MLIRLinalgStructuredOpsIncGenLib
+    MLIRMLProgramDialect
+    MLIRTensorDialect
+    iree::compiler::Dialect::LinalgExt::IR
+    iree::compiler::Dialect::Util::IR
+  PUBLIC
+)
+
+### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
