+++ compiler/src/iree/compiler/ExternalInterfaces/UtilExternalModels.h
@@ -4,17 +4,17 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#ifndef IREE_COMPILER_DIALECT_UTIL_IR_UTILEXTERNALMODELS_H_
-#define IREE_COMPILER_DIALECT_UTIL_IR_UTILEXTERNALMODELS_H_
+#ifndef IREE_COMPILER_EXTERNALINTERFACES_UTILEXTERNALMODELS_H_
+#define IREE_COMPILER_EXTERNALINTERFACES_UTILEXTERNALMODELS_H_
 
 namespace mlir {
 class DialectRegistry;
 } // namespace mlir
 
-namespace mlir::iree_compiler::IREE::Util {
+namespace mlir::iree_compiler {
 
 void registerUtilExternalModels(DialectRegistry &registry);
 
-} // namespace mlir::iree_compiler::IREE::Util
+} // namespace mlir::iree_compiler
 
-#endif // IREE_COMPILER_DIALECT_UTIL_IR_UTILEXTERNALMODELS_H_
+#endif // IREE_COMPILER_EXTERNALINTERFACES_UTILEXTERNALMODELS_H_
