+++ compiler/src/iree/compiler/Tools/BUILD.bazel
@@ -48,7 +48,6 @@ iree_compiler_cc_library(
         "//compiler/src/iree/compiler/Dialect/Stream/IR",
         "//compiler/src/iree/compiler/Dialect/Stream/Transforms",
         "//compiler/src/iree/compiler/Dialect/Util/IR",
-        "//compiler/src/iree/compiler/Dialect/Util/IR:ExternalModels",
         "//compiler/src/iree/compiler/Dialect/Util/TransformOps",
         "//compiler/src/iree/compiler/Dialect/Util/Transforms",
         "//compiler/src/iree/compiler/Dialect/VM/Analysis",
@@ -58,6 +57,7 @@ iree_compiler_cc_library(
         "//compiler/src/iree/compiler/Dialect/VMVX/IR:VMVXDialect",
         "//compiler/src/iree/compiler/Dialect/VMVX/Transforms",
         "//compiler/src/iree/compiler/Dialect/Vulkan/IR",
+        "//compiler/src/iree/compiler/ExternalInterfaces:ExternalModels",
         "//compiler/src/iree/compiler/GlobalOptimization/Interfaces",
         "//compiler/src/iree/compiler/InputConversion/Common",
         "//compiler/src/iree/compiler/Modules/HAL/Inline/IR:HALInlineDialect",
