+++ compiler/src/iree/compiler/Tools/CMakeLists.txt
@@ -52,7 +52,6 @@ iree_cc_library(
     iree::compiler::Dialect::Stream::IR
     iree::compiler::Dialect::Stream::Transforms
     iree::compiler::Dialect::Util::IR
-    iree::compiler::Dialect::Util::IR::ExternalModels
     iree::compiler::Dialect::Util::TransformOps
     iree::compiler::Dialect::Util::Transforms
     iree::compiler::Dialect::VM::Analysis
@@ -62,6 +61,7 @@ iree_cc_library(
     iree::compiler::Dialect::VMVX::IR::VMVXDialect
     iree::compiler::Dialect::VMVX::Transforms
     iree::compiler::Dialect::Vulkan::IR
+    iree::compiler::ExternalInterfaces::ExternalModels
     iree::compiler::GlobalOptimization::Interfaces::Interfaces
     iree::compiler::InputConversion::Common
     iree::compiler::Modules::HAL::Inline::IR::HALInlineDialect
